feat(render): implement buffered rendering

+++ include/chip8.h
@@ -39,5 +39,5 @@ private:
     uint8_t sound_timer{};
 
     std::mt19937 rand_gen;
-    std::uniform_int_distribution<uint8_t> rand_byte;
+    std::uniform_int_distribution<unsigned> rand_byte;
 };
